+++ lib/src/wallet/wallet.dart
@@ -1,3 +1,2 @@
 export 'hd_wallet.dart';
 export 'wallet_credentials.dart';
-export 'polygon_client.dart';